# lib/api-zod/src/generated/types/woundInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private wound-tracking and time-limited report-sharing API.
 * OpenAPI spec version: 0.1.0
 */

export interface WoundInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  bodySite: string;
}