# lib/api-zod/src/generated/types/woundShareInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Private wound-tracking and time-limited report-sharing API.
 * OpenAPI spec version: 0.1.0
 */
import type { WoundShareInputExpiresInHours } from './woundShareInputExpiresInHours';

export interface WoundShareInput {
  includeSymptoms: boolean;
  includeNotes: boolean;
  includeImages: boolean;
  expiresInHours: WoundShareInputExpiresInHours;
  /** @maxItems 100 */
  updateIds?: number[];
}